requirements.txt
streamlit==1.28.1
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
huggingface-hub==0.17.3
datasets==2.14.5
transformers==4.34.0
PyYAML==6.0.1
wandb==0.15.12